set up pcoip-agent to run from s6, remove selkies
This sets up the s6 structure to start HP anywhere.
It also removes Selkies (to avoid any conflicting behavior with Anywhere).

This is just a test setup. Once we'd confirm it as working, we'd get
this pushed to a separate repo for HP Anywhere, since it doesn't make
sense to run both it and Selkies.

Currently, upon setting that env and trying to start pcoip-agent, we
get:

```
helios  | System has not been booted with systemd as init system (PID 1). Can't operate.
helios  | Failed to connect to bus: Host is down

```